requirements.txt
numpy
pandas
matplotlib
notebook
ipykernel
python-dotenv
